requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
python-dotenv>=0.15.0
